feat: this adds support for adding custom transports
this change breaks the existing addTransport interface and refactors
it to addDefaultTransport. This will allow users to add a transport,
where they have access to the underlying protocol features.

The motivation behind this is to make it easy for people to extend
the transport to fit their usage context. This will also be useful
for people that need access to what the transport is doing like
receiving context specific events.

In the case of notifications coming from the server, this will
allow the developer to eventually hook into an event stream
and add handlers to their transport of choice, something that might
be tricker to do under the abstraction layer provided.

This change is a stepping stone to that.

fixes #619